Apply requested fixes to Easy GUI changes
