+++ frontend/Dockerstage
@@ -4,3 +4,4 @@ ARG OFRAK_DIR=.
 COPY --chown=node:node $OFRAK_DIR/frontend /home/node/frontend
 
 WORKDIR /home/node/frontend
+RUN su node -c "npm install && npm run build"
