+++ frontend/nginx.conf
@@ -5,14 +5,12 @@ server {
     client_max_body_size 2048M;
 
     root /ofrak_gui;
-    location / {}
+    location / {
+        proxy_pass http://127.0.0.1:8877/;
+    }
 
     location /docs/ {
         proxy_pass http://127.0.0.1:8000/;
-    }
-
-    location /api/ {
-        proxy_pass http://127.0.0.1:8877/;
     }
 
     # Read timeout currently needs to allow for basically all of an unpack_recursively call
