+++ ofrak-ghidra.yml
@@ -13,4 +13,4 @@ packages_paths:
 entrypoint: |
     nginx \
       & python3 -m ofrak_ghidra.server start \
-      & python3 /ofrak_server.py 0.0.0.0 8877
+      & python3 -m ofrak gui -H 0.0.0.0 -p 8877
